make separate file for fake curricula
